requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
jsonalias==0.1.1
multidict==6.0.5
nltk==3.8.1
numpy==2.0.1
oauthlib==3.2.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
solana==0.34.2
solders==0.21.0
soupsieve==2.6
textblob==0.18.0.post0
tornado==6.1
tqdm==4.66.4
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websockets==11.0.3
yarl==1.11.0
